Add BitNet manifold measurements, rank-reduced generation, and fixes
Stage 1 (BitNet b1.58-2B-4T, original ternary weights):
- Manifold mapped across all 31 layers
- TwoNN intrinsic dimensionality: ~10 constant across all layers
- PR profile: expand (10→143) → collapse (143→32), smooth relaxation
- Fixed bf16→fp16 overflow in cache (changed default dtype to bfloat16)
- Fixed non-finite value handling in SVD computations

Stage 4 new scripts:
- stage4_direct.py: measures per-token latency and KV attention entropy
  during generation without rank reduction (baseline characterization)
- stage4_rank_reduced.py: dynamic rank projection using SVD bases from
  Stage 1, with entropy-to-rank mapping during generation

Key findings:
- Manifold is ~10D regardless of layer depth (layers 4-30)
- PR follows expand→peak→collapse pattern (spin glass relaxation)
- KV entropy tracks token-level frustration/relaxation in real time
- CPU rank projection adds overhead (memory-bound); GPU needed for
  wall-clock speedup where FLOPs scale with rank
- Model changed to original BitNet (packed ternary, no fp16 overflow)

+++ src/measurement/cache_hidden_states.py
@@ -111,7 +111,7 @@ def cache_calibration_hidden_states(
     chunk_size: int = 512,
     batch_size: int = 2,
     iter_texts: Optional[Iterable[str]] = None,
-    dtype: torch.dtype = torch.float16,
+    dtype: torch.dtype = torch.bfloat16,
     source_name: str = "wikitext-103",
 ) -> CacheMeta:
     """Runs ``model`` on the calibration token stream and saves per-layer
